AutoSystem work + GearShift blank + FMS
Trying to flesh out the AutoSystem with actual code but don't have any
of the distances, will put distances in when I have the time to look
over the blueprints.
GearShift is empty and should be filled in order for drive team to get
what they want out of autonomous.
The results from FMS now appear in Robot as variable gameData as a
string. Also there are individual side string variables for the switch
and scale, if we want to put them on the driver station or something
like that.