The lift goes to a set distance
When the buttons x,b and y are clicked on operator control the robot lift will go to 4, 5 and 6 ft repectivley. The buttons only work when you want the lift to go up though so the lift can go from 4 to 5 but not back down that has to be done manually with the controller.