READ FULL DESCRIPTION - important
Added code for new pneumatics and networktables experiment
This should run the new pneumatics, though the port numbers may need to change. PLEASE CHANGE THE PORT NUMBERS TO MATCH THE COMMANDS, not vice versa. Everything should run well as long as this happens. doing it the other way may over complicate things and make it not work well.