Classpath must NEVER be touched or Indian will have your head...
So recently a commit was made that removed .classpath from the source files of our robot. This was attempted to be fixed, but, due to GitHub's merging system, was swiftly made a really big problem. Eclispe is responsible for the creation and monitering of ".classpath" This file is created once at the inception of the file and edits to it should not be made unless absolutely necesarry. The big issue lies with its removal. If removed, Eclispe, seeing that one was already created, assumes that you merely did not want anything in your classpath. This includes useful things like Java as a whole, WPILib, CTRE Pheonix, and other fun programs. I shouldn't have to say why those being missing is a problem. Anyways, to fix this issue I had to go into Eclipse's source code and make it think it never gave the project a classpath file, "Indiana Jones" our old one in (this being the one that already ignored which user it was looking at so as to avoid multi computer conflicts) and then make Eclipse think it had already, in fact, given us a classpath and refresh to accomodate said classpath file. Long story short, I used magic and fixed the problem. Don't do it again. :smirk: