Did Somebody Say Ramp?
The intent is to provide drivers with a sense of pride and accomplishment for unlocking different ramps. As for speed, we selected initial values based upon data from the Open Beta and other adjustments made to autonomous before launch. Among other things, we’re looking at average per-driver win rates on a daily basis, and we’ll be making constant adjustments to ensure that drivers have ramps that are compelling, rewarding, and of course attainable via gameplay. We appreciate the candid feedback, and the passion the community has put forth around the current topics here on GitHub, our pits and across numerous person to person outlets. Our team will continue to make changes and monitor community feedback and update everyone as soon and as often as we can.