I Could Have Sworn I Heard Ramp...
Did you ever hear the  Tragedy of Darth Nathaniel the Wise? I thought not. It's not a story the Software Devs would tell you. It's a Java legend. Nathaniel was a Dark Lord of Java so powerful and so wise, he could use the Code to influence the robot to create... autonomous. He had such a knowledge of the Code, he could even keep the programs he cared about... from having a ramp. The ramp of the Code is a pathway to many abilities some consider to be unnatural. He became so powerful, the only thing he was afraid of was losing his rampless code...which, eventually of course, he did. Unfortunately, he taught his apprentice everything he knew. Then his apprentice added a ramp in his sleep. Ironic. He could save others from the ramp... but not himself.