Second version of auto chooser, with conditionals
Figured out how to get if statements working in the auto chooser.  It works if it is done in autonomousInit and if you only use information that is available then.
Tested a simple auto routine with a L/R if statement and it works.
Now we need to plan how to use this.